+++ docs/reference/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - A new "--remote" argument is available for the deploy command to manage external executor config:
     sf-operator apply ./prod/sf.yaml --remote prod/executors/$hostname/sf.yaml
   It is no longer necessary to manually synchronize the zuul ssh key or the zuul connection secrets for example.
+- A new "sf-operator rotate-password" command line is available to perform service secret rotation.
 
 ### Changed